Revert "fix weigth to unsigned in class Direct_Solver_Context"
This reverts commit ac8f45f3e3d6cba45cb88fcc056763fe978c0dfe.

+++ src/metaSMT/DirectSolver_Context.hpp
@@ -54,7 +54,7 @@ namespace metaSMT {
         , Expr2 bw
     ) {
       const uint64_t val   = proto::value(value);
-      const unsigned width = proto::value(bw);
+      const uint64_t width = proto::value(bw);
 
       return SolverContext::operator() ( tag,
         boost::any(boost::make_tuple(val, width))
@@ -68,7 +68,7 @@ namespace metaSMT {
         , Expr2 bw
     ) {
       const int64_t val            = proto::value(value);
-      const unsigned width = proto::value(bw);
+      const uint64_t width = proto::value(bw);
 
       return  SolverContext::operator() ( tag, 
         boost::any(boost::make_tuple(val, width))
